requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
aiofiles==24.1.0
pdfplumber==0.11.4
python-docx==1.1.2
openai>=2.0.0
google-genai>=0.2.0
python-dotenv==1.0.1
